laundrytasker/requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.5
django-bootstrap3==10.0.1
django-braces==1.13.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
gunicorn==19.8.1
heroku==0.1.4
idna==2.6
oauthlib==2.1.0
Pillow==5.1.0
PyJWT==1.6.4
python-dateutil==1.5
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
stripe==2.0.3
urllib3==1.22
whitenoise==3.3.1